LRCI-7822 Extract BaseReport from BuildHistoryReport
